+++ evennia/server/evennia_launcher.py
@@ -13,6 +13,7 @@ Run the script with the -h flag to see usage information.
 
 import os
 import sys
+import re
 import signal
 import shutil
 import importlib
@@ -25,6 +26,7 @@ from subprocess import Popen, check_output, call, CalledProcessError, STDOUT
 from twisted.protocols import amp
 from twisted.internet import reactor, endpoints
 import django
+from django.core.management import execute_from_command_line
 
 # Signal processing
 SIG = signal.SIGINT
@@ -393,6 +395,7 @@ ERROR_DJANGO_MIN = \
     ERROR: Django {dversion} found. Evennia requires version {django_min}
     or higher.
 
+    TE_TEST
     If you are using a virtualenv, use the command `pip install --upgrade -e evennia` where
     `evennia` is the folder to where you cloned the Evennia library. If not
     in a virtualenv you can install django with for example `pip install --upgrade django`
@@ -428,16 +431,15 @@ NOTE_KEYBOARDINTERRUPT = \
 NOTE_TEST_DEFAULT = \
     """
     TESTING: Using Evennia's default settings file (evennia.settings_default).
-    (use 'evennia --settings settings.py test .' to run tests on the game dir)
+    (use 'evennia test --settings settings.py .' to run only your custom game tests)
     """
 
 NOTE_TEST_CUSTOM = \
     """
     TESTING: Using specified settings file '{settings_dotpath}'.
 
-    (Obs: Evennia's full test suite may not pass if the settings are very
-    different from the default. Use 'test .' as arguments to run only tests
-    on the game dir.)
+    OBS: Evennia's full test suite may not pass if the settings are very
+    different from the default (use 'evennia test evennia' to run core tests)
     """
 
 PROCESS_ERROR = \
@@ -519,7 +521,7 @@ def _print_info(portal_info_dict, server_info_dict):
         out = {}
         for key, value in dct.items():
             if isinstance(value, list):
-                value = "\n{}".format(ind).join(value)
+                value = "\n{}".format(ind).join(str(val) for val in value)
             out[key] = value
         return out
 
@@ -686,13 +688,14 @@ def send_instruction(operation, arguments, callback=None, errback=None):
 
     if AMP_CONNECTION:
         # already connected - send right away
-        _send()
+        return _send()
     else:
         # we must connect first, send once connected
         point = endpoints.TCP4ClientEndpoint(reactor, AMP_HOST, AMP_PORT)
         deferred = endpoints.connectProtocol(point, AMPLauncherProtocol())
         deferred.addCallbacks(_on_connect, _on_connect_fail)
         REACTOR_RUN = True
+        return deferred
 
 
 def query_status(callback=None):
@@ -2116,7 +2119,7 @@ def main():
             else:
                 kill(SERVER_PIDFILE, 'Server')
     elif option != "noop":
-        # pass-through to django manager
+        # pass-through to django manager, but set things up first
         check_db = False
         need_gamedir = True
         # some commands don't require the presence of a game directory to work
@@ -2136,31 +2139,16 @@ def main():
 
         init_game_directory(CURRENT_DIR, check_db=check_db, need_gamedir=need_gamedir)
 
-        # pass on to the manager
-        args = [option]
-        kwargs = {}
-        if unknown_args:
-            for arg in unknown_args:
-                if arg.startswith("--"):
-                    print("arg:", arg)
-                    if "=" in arg:
-                        arg, value = [p.strip() for p in arg.split("=", 1)]
-                    else:
-                        value = True
-                    kwargs[arg.lstrip("--")] = value
-                else:
-                    args.append(arg)
-
-        # makemessages needs a special syntax to not conflict with the -l option
-        if len(args) > 1 and args[0] == "makemessages":
-            args.insert(1, "-l")
-
-        try:
-            django.core.management.call_command(*args, **kwargs)
-        except django.core.management.base.CommandError as exc:
-            args = ", ".join(args)
-            kwargs = ", ".join(["--%s" % kw for kw in kwargs])
-            print(ERROR_INPUT.format(traceback=exc, args=args, kwargs=kwargs))
+        if option == "migrate":
+            # we have to launch migrate within the program to make sure migrations
+            # run within the scope of the launcher (otherwise missing a db will cause errors)
+            django.core.management.call_command(*([option] + unknown_args))
+        else:
+            # pass on to the core django manager - re-parse the entire input line
+            # but keep 'evennia' as the name instead of django-admin. This is
+            # an exit condition.
+            sys.argv[0] = re.sub(r'(-script\.pyw?|\.exe)?$', '', sys.argv[0])
+            sys.exit(execute_from_command_line())
 
     elif not args.tail_log:
         # no input; print evennia info (don't pring if we're tailing log)